[API] Update C-API semantics
Update C-API semantics about layer and optimizer
While the layer and optimizer are in use, do allow deletion of layer and optimizer
Further, once the layer and optimizer are set/added to a model, their ownership is transferred to the model
So, no need to delete them separately. Deleting the model will directly delete them as well.

Signed-off-by: Parichay Kapoor <pk.kapoor@samsung.com>