[optimizer] private ownership of optimizer in layer
This patch updates the ownership semantics of optimizer in the layer.
Before this patch, although, each layer creates its own optimizer, yet it was allowed to
be shared with a shared_ptr. With this patch, optimizer is a unique_ptr meaning
that it is owned by the layer and cannot be shared directly with other objects.

**Self evaluation:**
1. Build test: [x]Passed [ ]Failed [ ]Skipped
2. Run test: [x]Passed [ ]Failed [ ]Skipped

Signed-off-by: Parichay Kapoor <pk.kapoor@samsung.com>