[manager] Set the default usage timestamp for tensors
Set default usage time index for all the tensors based on when the layer
is to be run in the sorted list.
The usage time indexes are stored in each GraphNode by the GraphCore.
When creating the sorted list, each layer is updated to store its
order location/index when executed in this sequence for forwarding as
well as backwarding.

Signed-off-by: Parichay Kapoor <pk.kapoor@samsung.com>