[test/planner] Unittests for memory planner
This patch adds unittests for the memory planner.
The unittests covers some of the extreme scenarios for memory planning
and ensures that the constraints are satisfied.
Note that these tests do not test the efficient of the planning but the
validity of the planned layout.

Corresponding bug fixes are also added.

Signed-off-by: Parichay Kapoor <pk.kapoor@samsung.com>