[memorypool/test] Unittests for memory pool
This patch adds unittests for memory pool for its public APIs.
Further, memory validation is added for the requested memories and
those tests will be added for each planner. This ensures that the
memories returned by each planner do not overlap, and remain valid while
usage.

Fixes corresponding to the unittests are also added.

Signed-off-by: Parichay Kapoor <pk.kapoor@samsung.com>