[manager] Constraint input and ouput tensor sharing
Add the constraint of the output tensors of a node with the input
tensors of the next nodes. This allows memory pool to understand the
manager that these two inputs/outputs must use the same tensor and
enables tensor sharing.

This is acheieved by requesting already creating tensors with the given
name.

Signed-off-by: Parichay Kapoor <pk.kapoor@samsung.com>