[flatten] Remove in-place implementation
Flatten layer in-place implementation manipulates the tensor stored in
its input and output. This violates the execution order usage of the
tensors and does not work with the memory planning setup.
This patch makes flatten layer to work out-of-place with a copy.

There will soon be a patch to make flatten work in-place with proper
method.

Signed-off-by: Parichay Kapoor <pk.kapoor@samsung.com>